Completed edit password functionality.
*Completed edit password AJAX request and UI response.

**KNOWN ISSUE** The notification banner isn't fitting well with the
other elements in the carousel view. When I added it to the top, it
was hidden underneath the nav bar. I added it to the bottom of the
carousel view, actually caused it to overlay the nav bar at the top.
In any event, everything works, it's just a stylistic issue that we
will need to address.

**NEXT**
(1) Fix stylistic issue with Banner (maybe a clearfix issue or consider
another type of notification)
(2) Move on to images resource (starting with upload image)